backend/pyproject.toml
[project]
name = "gousto-meal-planner-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "greenlet>=3.1.1",
    "python-dotenv>=1.0.1",
]
